raise executor msg.value cap and add func to allow user configuration
When using TestHelper.sol to mock the LayerZero V2 environment for my tests, I was unable to pass more than 1 gwei to lzReceive() due to the limit imposed on the executor. I believe 1 gwei is far too low, so I raised it to 0.1 ether and added the setExecutorValueCap(uint128) function to TestHelper.sol so the user can adjust this value if their tests require it.